cleanup a few compilation warnings

+++ tinyHTTP/src/parsers/thttp_parser_url.c
@@ -504,7 +504,7 @@ _out: {
             56
             /* #line 125 "./ragel/thttp_parser_url.rl" */
       ) {
-        TSK_DEBUG_ERROR("Failed to parse HTTP/HTTPS URL: '%.*s'", length, urlstring);
+        TSK_DEBUG_ERROR("Failed to parse HTTP/HTTPS URL: '%.*s'", (int)length, urlstring);
         TSK_OBJECT_SAFE_FREE(url);
     }
     else if(!have_port) {
